pyproject.toml
[tool.poetry]
name = "aiosteady"
version = "0.1.0"
description = ""
authors = ["Tin Tvrtkovic <tinchester@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
attrs = "^20.3.0"
aioredis = "^1.3.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
black = {version = "^20.8b1", allow-prereleases = true}
pytest-asyncio = "^0.14.0"
mypy = "^0.812"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
